+++ index.js
@@ -325,5 +325,6 @@ export * from './src/components/modelHelper.js';
 export * from './src/components/composables.js';
 export { LegendType, StyleRowType } from './src/legend/legendHelper.js';
 export { default as Search } from './src/search/search.js';
+export { markText } from './src/search/markText.js';
 export { default as ResultItemComponent } from './src/search/ResultItem.vue';
 export { default as ResultsComponent } from './src/search/ResultsComponent.vue';